Add with-name function
